Fui incluída a funcionalidade para excluir aluno matriculado.

+++ src/main/java/br/leg/alrr/abrindocaminhos/controller/AlunoListagemMB.java
@@ -4,6 +4,7 @@ import br.leg.alrr.abrindocaminhos.model.Aluno;
 import br.leg.alrr.abrindocaminhos.model.Genitores;
 import br.leg.alrr.abrindocaminhos.model.Matricula;
 import br.leg.alrr.abrindocaminhos.persistence.AlunoDAO;
+import br.leg.alrr.abrindocaminhos.persistence.MatriculaDAO;
 import br.leg.alrr.abrindocaminhos.persistence.TurmaDAO;
 import br.leg.alrr.abrindocaminhos.util.DAOException;
 import br.leg.alrr.abrindocaminhos.util.FacesUtils;
@@ -30,17 +31,23 @@ public class AlunoListagemMB implements Serializable {
     @EJB
     private TurmaDAO turmaDAO;
 
+    @EJB
+    private MatriculaDAO matriculaDAO;
+
     private ArrayList<Aluno> alunos;
     private ArrayList<Aluno> irmaos;
     private ArrayList<Matricula> matriculas;
+    private ArrayList<Matricula> matriculasParaExcluir;
 
     private Aluno aluno;
     private Aluno irmao;
     private Genitores genitores;
 
     private boolean removerAluno = false;
+    private boolean removerAlunoComMatricula = false;
     private boolean mostrarTabelaDeCurso = false;
     private boolean exibirAtividade = false;
+    private boolean exibirModalAlunoMatricula = false;
 
     private Long idAluno;
     private Long idGenitor;
@@ -57,6 +64,21 @@ public class AlunoListagemMB implements Serializable {
                 aluno = (Aluno) FacesUtils.getBean("alunoAtividade");
                 FacesUtils.removeBean("alunoAtividade");
                 pegarTurma2();
+            }
+
+            if (FacesUtils.getBean("idAluno") != null) {
+                idAluno = (Long) FacesUtils.getBean("idAluno");
+                idEndereco = (Long) FacesUtils.getBean("idEndereco");
+                idGenitor = (Long) FacesUtils.getBean("idGenitor");
+                exibirModalAlunoMatricula = (boolean) FacesUtils.getBean("exibirModalAlunoMatricula");
+                matriculasParaExcluir = (ArrayList<Matricula>) FacesUtils.getBean("matriculasParaExcluir");
+
+                FacesUtils.removeBean("idAluno");
+                FacesUtils.removeBean("idEndereco");
+                FacesUtils.removeBean("idGenitor");
+                FacesUtils.removeBean("exibirModalAlunoMatricula");
+                FacesUtils.removeBean("matriculasParaExcluir");
+
             }
 
         } catch (Exception e) {
@@ -88,23 +110,72 @@ public class AlunoListagemMB implements Serializable {
         aluno = new Aluno();
     }
 
-    public void removerAluno() {
+    public String removerAluno() {
         try {
-            if (removerAluno) {
-                if (alunoDAO.genitorTemMaisDeUmFilho(new Genitores(idGenitor))) {
-                    alunoDAO.DeleteAluno(new Aluno(idAluno));
-                    alunoDAO.DeleteEndereco(idEndereco);
-                } else {
-                    alunoDAO.DeleteAluno(new Aluno(idAluno));
-                    alunoDAO.DeleteEndereco(idEndereco);
-                    alunoDAO.removerGenitor(alunoDAO.buscarGenitor(idGenitor));
-                }
-                FacesUtils.addInfoMessage("Aluno removido com sucesso!");
+
+            //PRIMEIRO VERIFICAR SE O ALUNO ESTÁMATRICULADO EM ALGUMA ATIVIDADE
+            if (!removerAlunoComMatricula) {
+                matriculasParaExcluir = (ArrayList<Matricula>) matriculaDAO.listarMatriculasPorAluno(new Aluno(idAluno));
+
+                FacesUtils.setBean("idAluno", idAluno);
+                FacesUtils.setBean("idEndereco", idEndereco);
+                FacesUtils.setBean("idGenitor", idGenitor);
+                FacesUtils.setBean("exibirModalAlunoMatricula", true);
+                FacesUtils.setBean("matriculasParaExcluir", matriculasParaExcluir);
+
             }
+
+            //O ALUNO NÃO TEM MATRÍCULA
+            if (matriculasParaExcluir.isEmpty()) {
+
+                if (removerAluno) {
+                    if (alunoDAO.genitorTemMaisDeUmFilho(new Genitores(idGenitor))) {
+                        alunoDAO.DeleteProntuario(idAluno);
+                        alunoDAO.DeleteAluno(new Aluno(idAluno));
+                        alunoDAO.DeleteEndereco(idEndereco);
+                    } else {
+                        alunoDAO.DeleteProntuario(idAluno);
+                        alunoDAO.DeleteAluno(new Aluno(idAluno));
+                        alunoDAO.DeleteEndereco(idEndereco);
+                        alunoDAO.removerGenitor(alunoDAO.buscarGenitor(idGenitor));
+                    }
+                    FacesUtils.addInfoMessageFlashScoped("Aluno removido com sucesso!");
+                    exibirModalAlunoMatricula = false;
+                }
+            } //O ALUNO TEM MATRICLA
+            else {
+
+                if (removerAlunoComMatricula) {
+
+                    if (alunoDAO.genitorTemMaisDeUmFilho(new Genitores(idGenitor))) {
+
+                        for (Matricula m : matriculasParaExcluir) {
+                            alunoDAO.DeleteMatricula(m.getId());
+                        }
+
+                        alunoDAO.DeleteProntuario(idAluno);
+                        alunoDAO.DeleteAluno(new Aluno(idAluno));
+                        alunoDAO.DeleteEndereco(idEndereco);
+                    } else {
+                        for (Matricula m : matriculasParaExcluir) {
+                            alunoDAO.DeleteMatricula(m.getId());
+                        }
+
+                        alunoDAO.DeleteProntuario(idAluno);
+                        alunoDAO.DeleteAluno(new Aluno(idAluno));
+                        alunoDAO.DeleteEndereco(idEndereco);
+                        alunoDAO.removerGenitor(alunoDAO.buscarGenitor(idGenitor));
+                    }
+                    FacesUtils.addInfoMessageFlashScoped("Aluno removido com sucesso!");
+                }
+                exibirModalAlunoMatricula = false;
+            }
+
         } catch (DAOException e) {
-            FacesUtils.addErrorMessage(e.getMessage());
+            FacesUtils.addErrorMessageFlashScoped(e.getMessage());
+            System.out.println(e.getCause());
         }
-        limparForm();
+        return "listar-editar-aluno.xhtml" + "?faces-redirect=true";
     }
 
     public void pegarTurma() {
@@ -117,7 +188,7 @@ public class AlunoListagemMB implements Serializable {
             FacesUtils.addErrorMessage(e.getMessage() + ": " + e.getCause());
         }
     }
-    
+
     public void pegarTurma2() {
         try {
             exibirAtividade = true;
@@ -152,8 +223,11 @@ public class AlunoListagemMB implements Serializable {
         aluno = new Aluno();
         genitores = new Genitores();
         alunos = new ArrayList<>();
+        matriculasParaExcluir = new ArrayList<>();
         mostrarTabelaDeCurso = false;
         removerAluno = false;
+        removerAlunoComMatricula = false;
+        exibirModalAlunoMatricula = false;
     }
 
     public void pesquisarFamilia() {
@@ -229,6 +303,26 @@ public class AlunoListagemMB implements Serializable {
         this.idAluno = idAluno;
         this.idGenitor = idGenitor;
         this.idEndereco = idEndereco;
+    }
+
+    public boolean isRemoverAlunoComMatricula() {
+        return removerAlunoComMatricula;
+    }
+
+    public void setRemoverAlunoComMatricula(boolean removerAlunoComMatricula) {
+        this.removerAlunoComMatricula = removerAlunoComMatricula;
+    }
+
+    public boolean isExibirModalAlunoMatricula() {
+        return exibirModalAlunoMatricula;
+    }
+
+    public void setExibirModalAlunoMatricula(boolean exibirModalAlunoMatricula) {
+        this.exibirModalAlunoMatricula = exibirModalAlunoMatricula;
+    }
+
+    public ArrayList<Matricula> getMatriculasParaExcluir() {
+        return matriculasParaExcluir;
     }
 
 }
